+++ CHANGELOG.md
@@ -7,3 +7,7 @@ Changes based off versioning in tags
 - Added EMA strategy and changes in class structure. 
 - Preparations for future addition of different Strategies
 
+# v3
+- Added Phemex support
+- Made the script job based 
+  - only single iteration which should be scheduled outside the application (cron / jobs)